move contactrestricted to team-building
