+++ shared/actions/fs/index.tsx
@@ -92,12 +92,14 @@ const loadAdditionalTlf = async (state: Container.TypedState, action: FsGen.Load
     )
   } catch (e) {
     if (e.code === RPCTypes.StatusCode.scteamcontactsettingsblock) {
-      const usernames = e.fields?.filter(elem => elem.key === 'usernames')
-      const requestee = usernames[0].value
+      const users = e.fields?.filter(elem => elem.key === 'usernames')
+      const usernames = users.map(elem => elem.value)
       // Don't leave the user on a broken FS dir screen.
       return [
         RouteTreeGen.createNavigateUp(),
-        RouteTreeGen.createNavigateAppend({path: [{props: {requestee}, selected: 'failedRequest'}]}),
+        RouteTreeGen.createNavigateAppend({
+          path: [{props: {source: 'newFolder', usernames}, selected: 'contactRestricted'}],
+        }),
       ]
     }
     return makeRetriableErrorHandler(action, action.payload.tlfPath)(e)
